==> 2_preparacao_dos_dados.Rmd <==
---
title: "Análise grandes desastres naturais"
author: "Mariana, Júlio, Lucas, Maria Lígia, Newton"
date: "02/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(dplyr)
library(readr)
library(janitor)
library(readODS)
library(stringi)  # stri_trans_general()
library(stringr)  # str_to_title()
library(rmarkdown)
```


## Importação

```{r}
dados <- readxl::read_xlsx("data/emdat_public_2022_05_02_query_uid-MrgSmc.xlsx",
                           skip = 6) %>% 
  clean_names()
```


## Variáveis

```{r}
names(dados)
```

## Glimpse para conferir se tipos das variáveis carregaram como esperado

```{r}
glimpse(dados)
```


## Frequência por Grupo de Desastre

```{r}
dados %>% 
  count(disaster_group) %>% 
  paged_table()
```


## Contagem de desastres por subgrupo

```{r}
dados %>% 
  count(disaster_subgroup, sort = T) %>% 
  paged_table()
```



## Frequência por Grupos, subgrupos, tipos, subtipos

```{r}
dados %>% 
  filter(disaster_group == "Natural") %>% 
  count(disaster_subgroup, disaster_type, disaster_subtype,
        disaster_subsubtype, sort = T) %>% 
  paged_table()
```




## Frequência por Grupos, subgrupos, tipos, subtipos

```{r}
dados %>% 
  filter(disaster_group == "Technological") %>% 
  count(disaster_subgroup, disaster_type, disaster_subtype, sort = T) %>% 
  paged_table()
```



## Filtro com foco no nosso objetivo

```{r}
dados <- dados %>% 
  filter(disaster_group == "Natural")
dados %>% 
  paged_table()
```


A partir daqui precisamos separar os eventos por unidades territoriais político-administrativas: municípios e estados


## Quebra da coluna de codificação em linhas

```{r}
dados <- dados %>% 
  tidyr::separate_rows(admin1_code, sep = ";") %>%
  tidyr::separate_rows(admin2_code, sep = ";") %>% 
  # Converte para funcionar join no próximo chunk
  mutate(admin1_code = as.integer(admin1_code),
         admin2_code = as.integer(admin2_code))

dados %>%  # Conferindo
  select(dis_no, adm_level, admin1_code, admin2_code) %>% 
  paged_table()
```


## Codificação GAUL

Rodar antes o script 1_junta_gaul1_gaul2.R

```{r}
gaul <- read_csv("data/gaul_br_adm1e2.csv", col_types = cols(gaul_adm1 = "i",
                                                             gaul_adm2 = "i"))

gaul_estados <- gaul %>% 
  distinct(gaul_adm1, adm1_name) %>% 
  filter(!is.na(gaul_adm1))  # Remove NA

gaul_municipios <- gaul %>% 
  # Junta rótulo de Município com de Estado para remover ambiguidade
  mutate(adm2_name = paste(adm2_name, "-", adm1_name)) %>% 
  select(gaul_adm2, adm2_name)

rm(gaul)
```


Junta os rótulos dos nomes das localidades usando os códigos Gaul

```{r}
dados <- dados %>% 
  left_join(gaul_estados, by = c("admin1_code" = "gaul_adm1")) %>% 
  left_join(gaul_municipios, by = c("admin2_code" = "gaul_adm2"))

dados %>% 
  select(geo_locations, admin1_code, adm1_name, admin2_code, adm2_name) %>% 
  paged_table()
```


## Codificação IBGE

Agora que temos os rótulos, podemos finalmente juntar o código IBGE que permitirá utilizar todos os indicadores que utilizam esta codificação.

https://www.ibge.gov.br/explica/codigos-dos-municipios.php

```{r}
ibge <- read_ods("data/RELATORIO_DTB_BRASIL_MUNICIPIO.ods") %>% 
  clean_names() %>% 
  # Selecionando o que interessa
  select(uf, nome_uf, municipio, codigo_municipio_completo, nome_municipio)
```


## Tabelas de equivalência Gaul - IBGE

Na base Gaul os rótulos dos Estados e Municípios estão no padrão ASCII (sem acentos gráficos) e os elementos de ligação têm inicial maiúscula, como "Rio De Janeiro" e "Abreu E Lima", etc. Então criamos tabelas de equivalência para posterior junção à base.

```{r}
ibge_estados <- ibge %>% 
  distinct(uf, nome_uf) %>% 
  mutate(nome_uf2 = stri_trans_general(nome_uf, "Latin-ASCII"),
         nome_uf2 = str_to_title(nome_uf2))

ibge_municipios <- ibge %>% 
  distinct(codigo_municipio_completo, nome_municipio, .keep_all = T) %>% 
  mutate(nome_municipio2 = stri_trans_general(nome_municipio, "Latin-ASCII"),
         nome_municipio2 = str_to_title(nome_municipio2),
         nome_uf2 = stri_trans_general(nome_uf, "Latin-ASCII"),
         nome_uf2 = str_to_title(nome_uf2),
         # Remove ambiguidade
         nome_municipio2 = paste(nome_municipio2, "-", nome_uf2)) %>% 
  select(codigo_municipio_completo, nome_municipio, nome_municipio2)

gaul_ibge_estados <- dados %>% 
  distinct(admin1_code, adm1_name) %>% 
  filter(!is.na(adm1_name)) %>% 
  arrange(adm1_name) %>% 
  left_join(ibge_estados, by = c("adm1_name" = "nome_uf2")) %>% 
  select(-adm1_name)  # Trabalharemos apenas com códigos
  

gaul_ibge_mun <- dados %>% 
  distinct(admin2_code, adm2_name) %>% 
  filter(!is.na(adm2_name)) %>% 
  arrange(adm2_name) %>% 
  left_join(ibge_municipios, by = c("adm2_name" = "nome_municipio2")) %>% 
  arrange(codigo_municipio_completo) %>% 
  select(-adm2_name)
  
```


## Último passo: juntar código IBGE na base de desastres

```{r}
dados <- dados %>% 
  left_join(gaul_ibge_estados, by = "admin1_code") %>% 
  left_join(gaul_ibge_mun, by = "admin2_code")

dados %>% 
  write_csv("data/desastres_por_uf_e_mun.csv")
```
